# doc/swagger.yml
openapi: 3.0.3
info:
  description: This is a sample PNL server
  version: 1.0.1
  title: PNL
servers:
  - url: http://localhost:8081/api
    description: Local server (uses live data)
paths:
...
  /v1/refreshtoken:
    post:
      tags:
        - auth
      summary: Validate RefreshToken
      description: Method allows to validate token
      operationId: refreshtoken
      requestBody:
        description: Capture object that needs to be validated in the system
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Token'
      responses:
        '200':
          description: Successful authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Login not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/admin/scenarios:
    get:
      tags:
        - scenarios
      summary: Get all scenarios for admin
      description: Method allows to get all scenarios
      operationId: getScenariosForAdmin
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArrayScenario'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - scenarios
      summary: Create scenario
      description: Method allows to create scenario
      operationId: postScenario
      requestBody:
        description: Request body to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Scenario'
      responses:
        '201':
          description: Successful
        '400':
          description: Invalid scenario struct supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Token:
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          description: access token
          example: ey1NiIsIn
        refreshToken:
          type: string
          description: refresh token
          example: pSjkuWmd
    Error:
      type: object
      description: Generic error response
      required:
        - data
      properties:
        data:
          type: object
          required:
            - code
          properties:
            code:
              type: integer
              format: int32
              description: The HTTP status code
              nullable: false
              example: '500'
            message:
              type: string
              description: The error message
              example: Internal server error
          example:
            data:
              code: 500
              message: Internal server error
    Scenario:
      type: object
      required:
        - ownerID
        - name
        - stepsCount
        - deletedAt
        - data
      properties:
        ownerID:
          type: integer
        name:
          type: string
          description: name scenario
        stepsCount:
          type: integer
          description: steps count in scenario
        deletedAt:
          type: string
          description: scenario deletion date
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScenarioData'
    ArrayScenario:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Scenario'
    ScenarioData:
      type: object
      required:
        - managers
        - projects
      properties:
        managers:
          type: array
          items:
            $ref: '#/components/schemas/Manager'
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    Manager:
      type: object
      required:
        - id
        - name
        - userID
        - employees
        - events
      properties:
        id:
          type: integer
        name:
          type: string
        userID:
          type: integer
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
    Project:
      type: object
      required:
        - id
        - name
        - rates
      properties:
        id:
          type: integer
        name:
          type: string
        rates:
          type: array
          items:
            $ref: '#/components/schemas/Rate'
    Rate:
      type: object
      required:
        - id
        - projectID
        - type
        - seniority
        - rate
        - illCompensation
        - vacationCompensation
      properties:
        id:
          type: integer
        projectID:
          type: integer
        type:
          type: string
          description: employee type (for example A1, A2, B1)
        seniority:
          type: integer
        rate:
          type: number
          description: rate per worker
        illCompensation:
          type: number
          description: sickness benefit
        vacationCompensation:
          type: number
          description: vacation pay
    Employee:
      type: object
      required:
        - id
        - name
        - seniority
        - salary
        - projectID
        - employeeStatus
      properties:
        id:
          type: integer
        name:
          type: string
        seniority:
          type: integer
        salary:
          type: number
        projectID:
          type: integer
        employeeStatus:
          type: string
          description: 'for example: on bench, vacation, ill'
    Event:
      type: object
      required:
        - id
        - eventTypeID
        - name
        - description
      properties:
        id:
          type: integer
        eventTypeID:
          type: integer
        name:
          type: string
        description:
          type: string
...